fix: component import in prod
