+++ components/index.tsx
@@ -1,4 +1,4 @@
-import SortableCondition from './react-sortable-condition'
-export { useTreeData } from './react-sortable-condition/useTreeData'
+import SortableCondition from './sortable-condition'
+export { useTreeData } from './sortable-condition/useTreeData'
 
-export default SortableCondition
+export { SortableCondition }
