--- components/react-sortable-condition/index.tsx
@@ -1,3 +1,0 @@
-import { SortableCondition } from './SortableCondition'
-
-export default SortableCondition
