+++ components/sortable-condition/Pattern.tsx
@@ -2,40 +2,30 @@ import React, { useContext, useMemo } from 'react'
 import cx from 'classnames'
 import isNull from 'lodash.isnull'
 
-import { NextPath, ConditionType, CustomConditionConfigs, ConditionItem } from './typings'
-import { ConfigContext } from './ConfigContext'
+import { NextPath, NormalType, CustomPatternConfigs, PatternItem } from './typings'
 import styles from './style/SortableCondition.css.json'
+import { ConfigContext } from './ConfigContext'
 import { DataContext } from './DataContext'
 import { isForbiddenConvert, isForbiddenCount } from './utils/rules'
-import './style/css'
 
-export type Props = {
-  type?: ConditionType
+type Props<T = any> = {
   path?: NextPath
-  node: ConditionItem
+  type: NormalType
+  node: PatternItem
+  patterns?: T
 }
 
-export const Condition = (props: Props) => {
-  const configs = useContext(ConfigContext).condition
+export const Pattern = (props: Props) => {
+  const configs = useContext(ConfigContext).pattern
   const globalConfigs = useContext(ConfigContext).global
-  const { treeData, dispatch } = useContext(DataContext)
-  const handleChangeConditionType = () => {
-    const nextType: ConditionType = props.type === 'and' ? 'or' : 'and'
-    dispatch({
-      type: 'CHANGE_TYPE',
-      payload: { path: props.path || [], type: nextType, node: props.node },
-    })
-    if (configs.onType) {
-      configs.onType(props.node, props.path || [], { type: nextType })
-    }
-  }
-  const handleAddCondition = () => {
+  const { dispatch, treeData } = useContext(DataContext)
+  const handleAddPattern = () => {
     dispatch({ type: 'ADD', payload: { path: props.path || [], node: props.node } })
     if (configs.onAdd) {
       configs.onAdd(props.node, props.path || [])
     }
   }
-  const handleDeleteCondition = () => {
+  const handleDeletePattern = () => {
     dispatch({
       type: 'DELETE',
       payload: { path: props.path || [], node: props.node },
@@ -44,7 +34,7 @@ export const Condition = (props: Props) => {
       configs.onDelete(props.node, props.path || [])
     }
   }
-  const handleConvertCondition = () => {
+  const handleConvert = () => {
     dispatch({
       type: 'CONVERT',
       payload: { path: props.path || [], node: props.node },
@@ -53,6 +43,18 @@ export const Condition = (props: Props) => {
       configs.onConvert(props.node, props.path || [])
     }
   }
+  const PatterComponent =
+    configs.component && React.isValidElement(configs.component)
+      ? React.cloneElement(configs.component, {
+          patterns: props.patterns,
+          onChange: ({ patterns }: { patterns: any }) => {
+            dispatch({
+              type: 'CHANGE_PATTERN',
+              payload: { path: props.path || [], patterns, node: props.node },
+            })
+          },
+        })
+      : 'this is pattern'
   const isNoConvertIcon =
     useMemo(() => isForbiddenConvert({ treeData, path: props.path, globalConfigs }), [
       treeData,
@@ -66,13 +68,13 @@ export const Condition = (props: Props) => {
   const isNoAddIcon = countStatus.add || isNull(configs.addIcon)
   const isNoDeleteIcon = countStatus.delete || isNull(configs.deleteIcon)
   return (
-    <div data-role="condition-item" className={cx(configs.className, styles.item)}>
-      <div data-role="content" onClick={handleChangeConditionType}>
-        <p>{props.type}</p>
+    <div data-role="pattern-item" className={cx(styles.item, configs.className)}>
+      <div data-role="content">
+        <p>{PatterComponent}</p>
       </div>
       <div data-role="btns" className={styles.btns}>
         {isNoConvertIcon ? null : (
-          <a data-role="convert-btn" className={styles.btn} onClick={handleConvertCondition}>
+          <a data-role="convert-btn" className={styles.btn} onClick={handleConvert}>
             {configs.convertIcon ? (
               configs.convertIcon
             ) : (
@@ -81,12 +83,12 @@ export const Condition = (props: Props) => {
           </a>
         )}
         {isNoAddIcon ? null : (
-          <a data-role="add-btn" className={styles.btn} onClick={handleAddCondition}>
+          <a data-role="add-btn" className={styles.btn} onClick={handleAddPattern}>
             {configs.addIcon ? configs.addIcon : <span className={styles.btn_content}>+</span>}
           </a>
         )}
         {isNoDeleteIcon ? null : (
-          <a data-role="delete-btn" className={styles.btn} onClick={handleDeleteCondition}>
+          <a data-role="delete-btn" className={styles.btn} onClick={handleDeletePattern}>
             {configs.deleteIcon ? (
               configs.deleteIcon
             ) : (
@@ -99,10 +101,10 @@ export const Condition = (props: Props) => {
   )
 }
 
-export type ConditionProps<T> = CustomConditionConfigs<T>
+export type PatternProps<T> = CustomPatternConfigs<T>
 
-export function ConfigCondition<T = any>(props: ConditionProps<T>) {
+export function ConfigPattern<T = any>(props: PatternProps<T>) {
   return <span>{props}</span>
 }
 
-ConfigCondition.displayName = 'Condition'
+ConfigPattern.displayName = 'Pattern'
