+++ components/sortable-condition/style/SortableCondition.css
@@ -1,0 +1,35 @@
+.SortableCondition_sortableCondition .rst__rowLabel {
+  width: 100%;
+  padding: 0px;
+}
+.SortableCondition_item {
+  position: relative;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  cursor: pointer;
+}
+.SortableCondition_item:hover .SortableCondition_btns {
+  opacity: 1;
+}
+.SortableCondition_btns {
+  position: absolute;
+  right: 0px;
+  transform: translateX(100%);
+  display: flex;
+  align-items: center;
+  padding-left: 8px;
+  opacity: 0;
+  cursor: pointer;
+}
+.SortableCondition_btn {
+  margin: 0 8px;
+}
+.SortableCondition_btn_content {
+  border: 1px solid #000;
+  display: block;
+  border-radius: 100%;
+  width: 21px;
+  text-align: center;
+}
