+++ components/sortable-condition/utils/extractConditionConfig.ts
