+++ components/sortable-condition/utils/extractPatternConfig.ts
