+++ components/sortable-condition/utils/getPatternsChangeTreeData.ts
