+++ components/sortable-condition/utils/getTypeChangeTreeData.ts
