+++ components/sortable-condition/utils/isAllConditionItems.ts
