+++ components/sortable-condition/utils/isAllNormalItems.ts
