+++ components/sortable-condition/utils/validate.ts
