+++ gulpfile.js
@@ -17,6 +17,7 @@ function compileStylus(modules) {
     .src(['components/**/*.styl'])
     .pipe(stylus())
     .pipe(postcss([config.postcssPlugins.autoprefixer, config.postcssPlugins.cssmodules]))
+    .pipe(gulp.dest(config.dirs.components))
     .pipe(gulp.dest(modules === false ? config.dirs.es : config.dirs.lib))
 }
